+++ ocfl-validate.py
@@ -61,7 +61,8 @@ for path in args.path:
         obj = ocfl.Object(lax_digests=args.lax_digests)
         if obj.validate_inventory(path,
                                   show_warnings=show_warnings,
-                                  show_errors=not args.very_quiet):
+                                  show_errors=not args.very_quiet,
+                                  extract_spec_version=True):
             num_good += 1
     else:
         log.error("Bad path %s (%s)", path, path_type)